feat: add esm function
